Move check live bill run service
We are going to need to always check if there is an existing live bill run before we create a new one, irrespective of whether it's supplementary or 2PT.

We do need to make a change to the service because it was hard coded to check for live supplementary bill runs only. Now we pass the billing batch type as an argument.
